Editing Comments is in the making.
Fixed comments not loading by changing the create method in ScenarioManager, it now takes a callback method as parameter and all things that were happening in the method were placed in the callback of loadTemplateScenario.
CommentMain is purely for the initialization of comment.html so it can run without external factors.